+++ quantify_core/analysis/single_qubit_timedomain.py
@@ -97,7 +97,10 @@ class SingleQubitTimedomainAnalysis(ba.BaseAnalysis):
         and if calibration points are present for the 0 and 1 state, populates
         self.dataset_processed.pop_exc with the excited state population.
         """
-        if self.dataset.y1.units == "deg":
+        if not hasattr(self.dataset, "y1"):
+            self.dataset_processed["S21"] = self.dataset.y0
+
+        elif self.dataset.y1.units == "deg":
             self.dataset_processed["S21"] = self.dataset.y0 * np.exp(
                 1j * np.deg2rad(self.dataset.y1)
             )
@@ -621,39 +624,63 @@ class AllXYAnalysis(SingleQubitTimedomainAnalysis):
         set_suptitle_from_dataset(fig, self.dataset)
 
 
-class RabiAnalysis(ba.BaseAnalysis):
+class RabiAnalysis(SingleQubitTimedomainAnalysis):
     """
     Fits a cosine curve to Rabi oscillation data and finds the qubit drive
     amplitude required to implement a pi-pulse.
+
+    The analysis will automatically rotate the data so that the data lies along the
+    axis with the best SNR.
     """
 
-    def process_data(self):
+    # pylint: disable=arguments-differ
+    def _rotate_to_calibrated_axis(self):
         """
-        Populates the :code:`.dataset_processed`.
+        If calibration points are True, automatically determine the point farthest
+        from the 0 point to use as a reference to rotate the data.
+
+        This will ensure the data lies along the axis with the best SNR.
         """
-        # y0 = amplitude, no check for the amplitude unit as the name/label is
-        # often different.
 
-        self.dataset_processed["Magnitude"] = self.dataset.y0
-        self.dataset_processed.Magnitude.attrs["name"] = "Magnitude"
-        self.dataset_processed.Magnitude.attrs["units"] = self.dataset.y0.units
-        self.dataset_processed.Magnitude.attrs["long_name"] = "Magnitude, $|S_{21}|$"
+        # index closest to rabi-pulse amplitude = 0
+        min_idx = abs(self.dataset_processed.x0.values).argmin()
+        # transmission measured for Rabi-pulse amplitude closest to 0
+        min_val = self.dataset_processed.S21[min_idx]
 
-        self.dataset_processed["x0"] = self.dataset.x0
-        self.dataset_processed = self.dataset_processed.set_coords("x0")
-        # replace the default dim_0 with x0
-        self.dataset_processed = self.dataset_processed.swap_dims({"dim_0": "x0"})
+        # find index with max absolute difference (distance in IQ space) to the S21_0
+        max_idx = abs(self.dataset_processed.S21 - min_val).argmax()
+        max_val = self.dataset_processed.S21[max_idx]
+
+        rotation_angle = np.angle(max_val - min_val)
+        rot_data = self.dataset_processed.S21 * np.exp(-1j * rotation_angle)
+
+        self.dataset_processed["S21_rot"] = rot_data
+        self.dataset_processed.S21_rot.attrs["name"] = "S21_rot"
+        self.dataset_processed.S21_rot.attrs["units"] = self.dataset.y0.units
+        self.dataset_processed.S21_rot.attrs[
+            "long_name"
+        ] = "Rotated transmission $S_{21}^R$"
+
+    def _choose_data_for_fit(self):
+        if self.calibration_points:
+            y_data = self.dataset_processed.S21_rot.real.values
+            x_data = self.dataset_processed.x0.values
+        else:
+            # if the data is not rotated,fit the magnitude of the signal
+            y_data = np.abs(self.dataset_processed.S21.values)
+            x_data = self.dataset_processed.x0.values
+
+        return x_data, y_data
 
     def run_fitting(self):
         """
         Fits a :class:`~quantify_core.analysis.fitting_models.RabiModel` to the data.
         """
         model = fm.RabiModel()
+        drive_amplitude, data = self._choose_data_for_fit()
 
-        magnitude = self.dataset_processed["Magnitude"].values
-        drive_amplitude = self.dataset_processed.x0.values
-        guess = model.guess(magnitude, drive_amp=drive_amplitude)
-        fit_result = model.fit(magnitude, params=guess, x=drive_amplitude)
+        guess = model.guess(data, drive_amp=drive_amplitude)
+        fit_result = model.fit(data, params=guess, x=drive_amplitude)
 
         self.fit_results.update({"Rabi_oscillation": fit_result})
 
@@ -705,7 +732,31 @@ class RabiAnalysis(ba.BaseAnalysis):
         # Add a textbox with the fit_message
         plot_textbox(ax, self.quantities_of_interest["fit_msg"])
 
-        self.dataset_processed.Magnitude.plot(ax=ax, marker=".", linestyle="")
+        if self.calibration_points:
+            ax.plot(
+                self.dataset_processed.x0,
+                self.dataset_processed.S21_rot.real,
+                marker=".",
+                linestyle="",
+            )
+
+            set_ylabel(
+                ax,
+                self.dataset_processed.S21_rot.long_name,
+                self.dataset_processed.S21_rot.units,
+            )
+        else:
+            ax.plot(
+                self.dataset_processed.x0,
+                self.dataset_processed.S21.real,
+                marker=".",
+                linestyle="",
+            )
+            set_ylabel(
+                ax,
+                self.dataset_processed.S21.long_name,
+                self.dataset_processed.S21.units,
+            )
 
         plot_fit(
             ax=ax,
@@ -714,9 +765,8 @@ class RabiAnalysis(ba.BaseAnalysis):
             range_casting="real",
         )
 
-        set_ylabel(ax, r"Output voltage", self.dataset_processed.Magnitude.units)
         set_xlabel(
             ax, self.dataset_processed.x0.long_name, self.dataset_processed.x0.units
         )
 
-        set_suptitle_from_dataset(fig, self.dataset, "S21")
+        set_suptitle_from_dataset(fig, self.dataset)