+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 Unreleased
 ----------
 
+* Analysis - Automatically rotate Rabi data to the axis with the best SNR (#249, !223)
 * Analysis - Added support for calibration points to rotate and scale data to a calibrated axis for single-qubit timedomain experiments (T1, Echo, Ramsey and AllXY) (#227,  !219)
 * Analysis - Added extra constraints to fits for T1, Echo and Ramsey when using a calibrated axis (T1, Echo, Ramsey) (#236,  !219)
 * Bugfix - Solved a bug where a fit would fail for a Ramsey experiment with negative values (#246, !219)